update dependencies (LND v0.18.0-beta and co)
Updated versions:

  - lnd: v0.17.4-beta -> v0.18.0-beta
  - lnd/kvdb: v1.4.4 -> v1.4.8
  - lnd/tor: v1.1.2 -> v1.1.3
  - lnd/healthcheck: v1.2.3 -> v1.2.4
  - lnd/tlv: v1.1.1 -> v1.2.6
  - neutrino: v0.16.0 -> v0.16.1-0.20240425105051-602843d34ffd
  - neutrino/cache: v1.1.1 -> v1.1.2
  - lndclient: v0.17.4-1 -> v0.18.0-2
  - loop: v0.26.6-beta -> v0.28.3
  - loop/swapserverrpc: v1.0.5 -> v1.0.8
  - pool: v0.6.2-beta.0.20230329135228-c3bffb52df3a ->
        v0.6.5-beta.0.20240531084722-4000ec802aaa
  - pool/auctioneerrpc: v1.0.7 -> v1.1.2
  - aperture: v0.1.21-beta.0.20230705004936-87bb996a4030 -> v0.3.2-beta
  - replace of protobuf: v1.30.0-hex-display -> v1.33.0-hex-display
  - Go: 1.21 -> 1.22.3

Pool version v0.6.5-beta.0.20240531084722-4000ec802aaa was used because it uses
LND v0.18.0-beta, not v0.18.0-beta.rc3.

Fixed imports and API changes.

Updated Go version in README.

Updated LND version in cmd/chantools/root.go.
Ran `make docs`.

+++ cmd/chantools/closepoolaccount.go
@@ -292,7 +292,7 @@ func closePoolAccount(extendedKey *hdkeychain.ExtendedKey, apiURL string,
 		signDesc.SignMethod = input.TaprootScriptSpendSignMethod
 	}
 	feeRateKWeight := chainfee.SatPerKVByte(1000 * feeRate).FeePerKWeight()
-	totalFee := feeRateKWeight.FeeForWeight(int64(estimator.Weight()))
+	totalFee := feeRateKWeight.FeeForWeight(estimator.Weight())
 
 	// Add our sweep destination output.
 	sweepTx.TxOut = []*wire.TxOut{{
